refactor login ui
